+++ src/Filters/LikeFilter.php
@@ -3,8 +3,6 @@ declare(strict_types=1);
 
 namespace Level23\Druid\Filters;
 
-use Level23\Druid\Extractions\ExtractionInterface;
-
 /**
  * Class LikeFilter
  *
@@ -21,8 +19,6 @@ class LikeFilter implements FilterInterface
 
     protected string $escapeCharacter;
 
-    protected ?ExtractionInterface $extractionFunction;
-
     /**
      * LikeFilter constructor.
      *
@@ -30,18 +26,15 @@ class LikeFilter implements FilterInterface
      * @param string                   $pattern                 LIKE pattern, such as "foo%" or "___bar".
      * @param string                   $escapeCharacter         An escape character that can be used to escape special
      *                                                          characters.
-     * @param ExtractionInterface|null $extractionFunction
      */
     public function __construct(
         string $dimension,
         string $pattern,
-        string $escapeCharacter = '\\',
-        ?ExtractionInterface $extractionFunction = null
+        string $escapeCharacter = '\\'
     ) {
-        $this->dimension          = $dimension;
-        $this->pattern            = $pattern;
-        $this->escapeCharacter    = $escapeCharacter;
-        $this->extractionFunction = $extractionFunction;
+        $this->dimension       = $dimension;
+        $this->pattern         = $pattern;
+        $this->escapeCharacter = $escapeCharacter;
     }
 
     /**
@@ -51,17 +44,11 @@ class LikeFilter implements FilterInterface
      */
     public function toArray(): array
     {
-        $result = [
+        return [
             'type'      => 'like',
             'dimension' => $this->dimension,
             'pattern'   => $this->pattern,
             'escape'    => $this->escapeCharacter,
         ];
-
-        if ($this->extractionFunction) {
-            $result['extractionFn'] = $this->extractionFunction->toArray();
-        }
-
-        return $result;
     }
 }