+++ tests/Filters/LikeFilterTest.php
@@ -5,21 +5,11 @@ namespace Level23\Druid\Tests\Filters;
 
 use Level23\Druid\Tests\TestCase;
 use Level23\Druid\Filters\LikeFilter;
-use Level23\Druid\Extractions\LookupExtraction;
 
 class LikeFilterTest extends TestCase
 {
-    /**
-     * @param bool $useExtractionFunction
-     * @testWith [true]
-     *           [false]
-     */
-    public function testFilter(bool $useExtractionFunction): void
+    public function testFilter(): void
     {
-        $extractionFunction = new LookupExtraction(
-            'full_username', false
-        );
-
         $expected = [
             'type'      => 'like',
             'dimension' => 'name',
@@ -27,12 +17,7 @@ class LikeFilterTest extends TestCase
             'escape'    => '#',
         ];
 
-        if ($useExtractionFunction) {
-            $filter                   = new LikeFilter('name', 'D%', '#', $extractionFunction);
-            $expected['extractionFn'] = $extractionFunction->toArray();
-        } else {
-            $filter = new LikeFilter('name', 'D%', '#');
-        }
+        $filter = new LikeFilter('name', 'D%', '#');
 
         $this->assertEquals($expected, $filter->toArray());
     }