+++ tests/Filters/RegexFilterTest.php
@@ -5,33 +5,18 @@ namespace Level23\Druid\Tests\Filters;
 
 use Level23\Druid\Tests\TestCase;
 use Level23\Druid\Filters\RegexFilter;
-use Level23\Druid\Extractions\LookupExtraction;
 
 class RegexFilterTest extends TestCase
 {
-    /**
-     * @param bool $useExtractionFunction
-     * @testWith [true]
-     *           [false]
-     */
-    public function testFilter(bool $useExtractionFunction): void
+    public function testFilter(): void
     {
-        $extractionFunction = new LookupExtraction(
-            'full_username', false
-        );
-
         $expected = [
             'type'      => 'regex',
             'dimension' => 'name',
             'pattern'   => '^[a-z]*$',
         ];
 
-        if ($useExtractionFunction) {
-            $filter                   = new RegexFilter('name', '^[a-z]*$', $extractionFunction);
-            $expected['extractionFn'] = $extractionFunction->toArray();
-        } else {
-            $filter = new RegexFilter('name', '^[a-z]*$');
-        }
+        $filter = new RegexFilter('name', '^[a-z]*$');
 
         $this->assertEquals($expected, $filter->toArray());
     }